api/requirements.txt
tensorflow==2.16.1
tensorflow-cpu==2.16.1
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy